Added arrowkey movement support, fixed general movement
- speed now working correct
- fixed flickering of usershape
- added MultiKeyPressListener to handle multiple key presses at once

+++ src/net/teammagic/mpgame/Player.java
@@ -3,7 +3,7 @@ package net.teammagic.mpgame;
 import java.awt.geom.Rectangle2D;
 
 public class Player extends Rectangle2D.Double {
-    int speed = 5;
+    double speed = 5;
 
     public Player(double x, double y, double w, double h) {
         super(x, y, w, h);
